Create internal directory and move headers to it
